Implemented routing according to the spec
Unauthenticated users can only access the landing page

Authenticated users cannot access the landing page
