+++ client/src/App.tsx
@@ -3,19 +3,12 @@ import { useDispatch, useSelector } from 'react-redux';
 import { Redirect, Route, Switch } from 'react-router-dom';
 
 import Loading from './pages/Loading/Loading.page';
+import LandingPage from './pages/LandingPage/LandingPage.component';
 import { fetchCurrentUser } from './redux/user/user.actions';
 import { RootState } from './redux/root.reducer';
 import './App.css';
 
-const NotFound = React.lazy(() => import('./pages/NotFound/NotFound.page'));
-const CGAppBar = React.lazy(() => import('./components/CGAppBar/CGAppBar.component'));
-const CreateProjectDialog = React.lazy(() =>
-	import('./components/CreateProjectDialog/CreateProjectDialog.component')
-);
-const LandingPage = React.lazy(() => import('./pages/LandingPage/LandingPage.component'));
-
-const ProjectHub = React.lazy(() => import('./pages/ProjectHub/ProjectHub.page'));
-const Dashboard = React.lazy(() => import('./pages/Dashboard/Dashboard.component'));
+const Authenticated = React.lazy(() => import('./Authenticated'));
 
 const App: React.FC = () => {
 	const isAuthenticated = useSelector((state: RootState) => state.user !== null);
@@ -28,26 +21,18 @@ const App: React.FC = () => {
 
 	return (
 		<div className='App'>
-			<Suspense fallback={<Loading />}>
-				{isAuthenticated ? (
-					<>
-						<CGAppBar />
-						<CreateProjectDialog />
-					</>
-				) : (
+			{isAuthenticated ? (
+				<Suspense fallback={<Loading />}>
+					<Authenticated />
+				</Suspense>
+			) : (
+				<>
 					<Redirect to='/' />
-				)}
-
-				<Switch>
-					<Route exact path='/'>
-						{isAuthenticated ? <Redirect to='/me' /> : <LandingPage />}
-					</Route>
-					<Route exact path='/me' component={Dashboard} />
-					<Route exact path='/loading' component={Loading} />
-					<Route path='/projects/:id' component={ProjectHub} />
-					<Route path='*' component={NotFound} />
-				</Switch>
-			</Suspense>
+					<Switch>
+						<Route exact path='/' component={LandingPage} />
+					</Switch>
+				</>
+			)}
 		</div>
 	);
 };
